split twitter
